# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BPMS API",
        "description": "API documentation for BPMS app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dots-approval/multiple": {
            "post": {
                "tags": [
                    "Views For Employee"
                ],
                "summary": "Get approval data for multiple cost centers",
                "operationId": "eae11b093a60f0256bb22faa7e24779c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cost_centers"
                                ],
                                "properties": {
                                    "cost_centers": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "CC001"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "cost_center": {
                                                        "type": "string",
                                                        "example": "CC001"
                                                    },
                                                    "cost_center_name": {
                                                        "type": "string",
                                                        "example": "Finance Department"
                                                    },
                                                    "cost_center_apv_1": {
                                                        "type": "string",
                                                        "example": "APV1"
                                                    },
                                                    "cost_center_apv_2": {
                                                        "type": "string",
                                                        "example": "APV2"
                                                    },
                                                    "BP": {
                                                        "type": "string",
                                                        "example": "BP001"
                                                    },
                                                    "employee_name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "role_user": {
                                                        "type": "string",
                                                        "example": "DH"
                                                    },
                                                    "title_name": {
                                                        "type": "string",
                                                        "example": "Finance Manager"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john.doe@company.com"
                                                    },
                                                    "sex": {
                                                        "type": "string",
                                                        "example": "MALE"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}